+++ runners/flink/src/test/java/org/apache/beam/runners/flink/FlinkExecutionEnvironmentsTest.java
@@ -29,6 +29,8 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.net.InetSocketAddress;
 import java.nio.file.Files;
+import java.util.Arrays;
+import java.util.Collection;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
@@ -49,17 +51,33 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
 import org.junit.rules.TemporaryFolder;
+import org.junit.runner.RunWith;
+import org.junit.runners.Parameterized;
 import org.powermock.reflect.Whitebox;
 
 /** Tests for {@link FlinkExecutionEnvironments}. */
+@RunWith(Parameterized.class)
 public class FlinkExecutionEnvironmentsTest {
 
   @Rule public TemporaryFolder temporaryFolder = new TemporaryFolder();
   @Rule public ExpectedException expectedException = ExpectedException.none();
 
+  @Parameterized.Parameter public boolean useDataStreamForBatch;
+
+  @Parameterized.Parameters(name = "UseDataStreamForBatch = {0}")
+  public static Collection<Object[]> useDataStreamForBatchJobValues() {
+    return Arrays.asList(new Object[][] {{false}, {true}});
+  }
+
+  private FlinkPipelineOptions getDefaultPipelineOptions() {
+    FlinkPipelineOptions options = FlinkPipelineOptions.defaults();
+    options.setUseDataStreamForBatch(useDataStreamForBatch);
+    return options;
+  }
+
   @Test
   public void shouldSetParallelismBatch() {
-    FlinkPipelineOptions options = FlinkPipelineOptions.defaults();
+    FlinkPipelineOptions options = getDefaultPipelineOptions();
     options.setRunner(TestFlinkRunner.class);
     options.setParallelism(42);
 
@@ -71,7 +89,7 @@ public class FlinkExecutionEnvironmentsTest {
 
   @Test
   public void shouldSetParallelismStreaming() {
-    FlinkPipelineOptions options = FlinkPipelineOptions.defaults();
+    FlinkPipelineOptions options = getDefaultPipelineOptions();
     options.setRunner(TestFlinkRunner.class);
     options.setParallelism(42);
 
@@ -84,7 +102,7 @@ public class FlinkExecutionEnvironmentsTest {
 
   @Test
   public void shouldSetMaxParallelismStreaming() {
-    FlinkPipelineOptions options = FlinkPipelineOptions.defaults();
+    FlinkPipelineOptions options = getDefaultPipelineOptions();
     options.setRunner(TestFlinkRunner.class);
     options.setMaxParallelism(42);
 
@@ -99,7 +117,7 @@ public class FlinkExecutionEnvironmentsTest {
   public void shouldInferParallelismFromEnvironmentBatch() throws IOException {
     String flinkConfDir = extractFlinkConfig();
 
-    FlinkPipelineOptions options = FlinkPipelineOptions.defaults();
+    FlinkPipelineOptions options = getDefaultPipelineOptions();
     options.setRunner(TestFlinkRunner.class);
     options.setFlinkMaster("host:80");
 
@@ -115,7 +133,7 @@ public class FlinkExecutionEnvironmentsTest {
   public void shouldInferParallelismFromEnvironmentStreaming() throws IOException {
     String confDir = extractFlinkConfig();
 
-    FlinkPipelineOptions options = FlinkPipelineOptions.defaults();
+    FlinkPipelineOptions options = getDefaultPipelineOptions();
     options.setRunner(TestFlinkRunner.class);
     options.setFlinkMaster("host:80");
 
@@ -129,7 +147,7 @@ public class FlinkExecutionEnvironmentsTest {
 
   @Test
   public void shouldFallbackToDefaultParallelismBatch() {
-    FlinkPipelineOptions options = FlinkPipelineOptions.defaults();
+    FlinkPipelineOptions options = getDefaultPipelineOptions();
     options.setRunner(TestFlinkRunner.class);
     options.setFlinkMaster("host:80");
 
@@ -141,7 +159,7 @@ public class FlinkExecutionEnvironmentsTest {
 
   @Test
   public void shouldFallbackToDefaultParallelismStreaming() {
-    FlinkPipelineOptions options = FlinkPipelineOptions.defaults();
+    FlinkPipelineOptions options = getDefaultPipelineOptions();
     options.setRunner(TestFlinkRunner.class);
     options.setFlinkMaster("host:80");
 
@@ -154,7 +172,7 @@ public class FlinkExecutionEnvironmentsTest {
 
   @Test
   public void useDefaultParallelismFromContextBatch() {
-    FlinkPipelineOptions options = FlinkPipelineOptions.defaults();
+    FlinkPipelineOptions options = getDefaultPipelineOptions();
     options.setRunner(TestFlinkRunner.class);
 
     ExecutionEnvironment bev = FlinkExecutionEnvironments.createBatchExecutionEnvironment(options);
@@ -166,7 +184,7 @@ public class FlinkExecutionEnvironmentsTest {
 
   @Test
   public void useDefaultParallelismFromContextStreaming() {
-    FlinkPipelineOptions options = FlinkPipelineOptions.defaults();
+    FlinkPipelineOptions options = getDefaultPipelineOptions();
     options.setRunner(TestFlinkRunner.class);
 
     StreamExecutionEnvironment sev =
@@ -179,7 +197,7 @@ public class FlinkExecutionEnvironmentsTest {
 
   @Test
   public void shouldParsePortForRemoteEnvironmentBatch() {
-    FlinkPipelineOptions options = FlinkPipelineOptions.defaults();
+    FlinkPipelineOptions options = getDefaultPipelineOptions();
     options.setRunner(FlinkRunner.class);
     options.setFlinkMaster("host:1234");
 
@@ -191,7 +209,7 @@ public class FlinkExecutionEnvironmentsTest {
 
   @Test
   public void shouldParsePortForRemoteEnvironmentStreaming() {
-    FlinkPipelineOptions options = FlinkPipelineOptions.defaults();
+    FlinkPipelineOptions options = getDefaultPipelineOptions();
     options.setRunner(FlinkRunner.class);
     options.setFlinkMaster("host:1234");
 
@@ -204,7 +222,7 @@ public class FlinkExecutionEnvironmentsTest {
 
   @Test
   public void shouldAllowPortOmissionForRemoteEnvironmentBatch() {
-    FlinkPipelineOptions options = FlinkPipelineOptions.defaults();
+    FlinkPipelineOptions options = getDefaultPipelineOptions();
     options.setRunner(FlinkRunner.class);
     options.setFlinkMaster("host");
 
@@ -216,7 +234,7 @@ public class FlinkExecutionEnvironmentsTest {
 
   @Test
   public void shouldAllowPortOmissionForRemoteEnvironmentStreaming() {
-    FlinkPipelineOptions options = FlinkPipelineOptions.defaults();
+    FlinkPipelineOptions options = getDefaultPipelineOptions();
     options.setRunner(FlinkRunner.class);
     options.setFlinkMaster("host");
 
@@ -229,7 +247,7 @@ public class FlinkExecutionEnvironmentsTest {
 
   @Test
   public void shouldTreatAutoAndEmptyHostTheSameBatch() {
-    FlinkPipelineOptions options = FlinkPipelineOptions.defaults();
+    FlinkPipelineOptions options = getDefaultPipelineOptions();
     options.setRunner(FlinkRunner.class);
 
     ExecutionEnvironment sev = FlinkExecutionEnvironments.createBatchExecutionEnvironment(options);
@@ -243,7 +261,7 @@ public class FlinkExecutionEnvironmentsTest {
 
   @Test
   public void shouldTreatAutoAndEmptyHostTheSameStreaming() {
-    FlinkPipelineOptions options = FlinkPipelineOptions.defaults();
+    FlinkPipelineOptions options = getDefaultPipelineOptions();
     options.setRunner(FlinkRunner.class);
 
     StreamExecutionEnvironment sev =
@@ -259,7 +277,7 @@ public class FlinkExecutionEnvironmentsTest {
 
   @Test
   public void shouldDetectMalformedPortBatch() {
-    FlinkPipelineOptions options = FlinkPipelineOptions.defaults();
+    FlinkPipelineOptions options = getDefaultPipelineOptions();
     options.setRunner(FlinkRunner.class);
     options.setFlinkMaster("host:p0rt");
 
@@ -271,7 +289,7 @@ public class FlinkExecutionEnvironmentsTest {
 
   @Test
   public void shouldDetectMalformedPortStreaming() {
-    FlinkPipelineOptions options = FlinkPipelineOptions.defaults();
+    FlinkPipelineOptions options = getDefaultPipelineOptions();
     options.setRunner(FlinkRunner.class);
     options.setFlinkMaster("host:p0rt");
 
@@ -283,7 +301,7 @@ public class FlinkExecutionEnvironmentsTest {
 
   @Test
   public void shouldSupportIPv4Batch() {
-    FlinkPipelineOptions options = FlinkPipelineOptions.defaults();
+    FlinkPipelineOptions options = getDefaultPipelineOptions();
     options.setRunner(FlinkRunner.class);
 
     options.setFlinkMaster("192.168.1.1:1234");
@@ -297,7 +315,7 @@ public class FlinkExecutionEnvironmentsTest {
 
   @Test
   public void shouldSupportIPv4Streaming() {
-    FlinkPipelineOptions options = FlinkPipelineOptions.defaults();
+    FlinkPipelineOptions options = getDefaultPipelineOptions();
     options.setRunner(FlinkRunner.class);
 
     options.setFlinkMaster("192.168.1.1:1234");
@@ -311,7 +329,7 @@ public class FlinkExecutionEnvironmentsTest {
 
   @Test
   public void shouldSupportIPv6Batch() {
-    FlinkPipelineOptions options = FlinkPipelineOptions.defaults();
+    FlinkPipelineOptions options = getDefaultPipelineOptions();
     options.setRunner(FlinkRunner.class);
 
     options.setFlinkMaster("[FE80:CD00:0000:0CDE:1257:0000:211E:729C]:1234");
@@ -326,7 +344,7 @@ public class FlinkExecutionEnvironmentsTest {
 
   @Test
   public void shouldSupportIPv6Streaming() {
-    FlinkPipelineOptions options = FlinkPipelineOptions.defaults();
+    FlinkPipelineOptions options = getDefaultPipelineOptions();
     options.setRunner(FlinkRunner.class);
 
     options.setFlinkMaster("[FE80:CD00:0000:0CDE:1257:0000:211E:729C]:1234");
@@ -342,7 +360,7 @@ public class FlinkExecutionEnvironmentsTest {
 
   @Test
   public void shouldRemoveHttpProtocolFromHostBatch() {
-    FlinkPipelineOptions options = FlinkPipelineOptions.defaults();
+    FlinkPipelineOptions options = getDefaultPipelineOptions();
     options.setRunner(FlinkRunner.class);
 
     for (String flinkMaster :
@@ -358,7 +376,7 @@ public class FlinkExecutionEnvironmentsTest {
 
   @Test
   public void shouldRemoveHttpProtocolFromHostStreaming() {
-    FlinkPipelineOptions options = FlinkPipelineOptions.defaults();
+    FlinkPipelineOptions options = getDefaultPipelineOptions();
     options.setRunner(FlinkRunner.class);
 
     for (String flinkMaster :
@@ -382,7 +400,7 @@ public class FlinkExecutionEnvironmentsTest {
   @Test
   public void shouldAutoSetIdleSourcesFlagWithoutCheckpointing() {
     // Checkpointing disabled, shut down sources immediately
-    FlinkPipelineOptions options = FlinkPipelineOptions.defaults();
+    FlinkPipelineOptions options = getDefaultPipelineOptions();
     FlinkExecutionEnvironments.createStreamExecutionEnvironment(options);
     assertThat(options.getShutdownSourcesAfterIdleMs(), is(0L));
   }
@@ -390,7 +408,7 @@ public class FlinkExecutionEnvironmentsTest {
   @Test
   public void shouldAutoSetIdleSourcesFlagWithCheckpointing() {
     // Checkpointing is enabled, never shut down sources
-    FlinkPipelineOptions options = FlinkPipelineOptions.defaults();
+    FlinkPipelineOptions options = getDefaultPipelineOptions();
     options.setCheckpointingInterval(1000L);
     FlinkExecutionEnvironments.createStreamExecutionEnvironment(options);
     assertThat(options.getShutdownSourcesAfterIdleMs(), is(Long.MAX_VALUE));
@@ -399,7 +417,7 @@ public class FlinkExecutionEnvironmentsTest {
   @Test
   public void shouldAcceptExplicitlySetIdleSourcesFlagWithoutCheckpointing() {
     // Checkpointing disabled, accept flag
-    FlinkPipelineOptions options = FlinkPipelineOptions.defaults();
+    FlinkPipelineOptions options = getDefaultPipelineOptions();
     options.setShutdownSourcesAfterIdleMs(42L);
     FlinkExecutionEnvironments.createStreamExecutionEnvironment(options);
     assertThat(options.getShutdownSourcesAfterIdleMs(), is(42L));
@@ -408,7 +426,7 @@ public class FlinkExecutionEnvironmentsTest {
   @Test
   public void shouldAcceptExplicitlySetIdleSourcesFlagWithCheckpointing() {
     // Checkpointing enable, still accept flag
-    FlinkPipelineOptions options = FlinkPipelineOptions.defaults();
+    FlinkPipelineOptions options = getDefaultPipelineOptions();
     options.setCheckpointingInterval(1000L);
     options.setShutdownSourcesAfterIdleMs(42L);
     FlinkExecutionEnvironments.createStreamExecutionEnvironment(options);
@@ -418,7 +436,7 @@ public class FlinkExecutionEnvironmentsTest {
   @Test
   public void shouldSetSavepointRestoreForRemoteStreaming() {
     String path = "fakePath";
-    FlinkPipelineOptions options = FlinkPipelineOptions.defaults();
+    FlinkPipelineOptions options = getDefaultPipelineOptions();
     options.setRunner(TestFlinkRunner.class);
     options.setFlinkMaster("host:80");
     options.setSavepointPath(path);
@@ -432,7 +450,7 @@ public class FlinkExecutionEnvironmentsTest {
 
   @Test
   public void shouldFailOnUnknownStateBackend() {
-    FlinkPipelineOptions options = FlinkPipelineOptions.defaults();
+    FlinkPipelineOptions options = getDefaultPipelineOptions();
     options.setStreaming(true);
     options.setStateBackend("unknown");
     options.setStateBackendStoragePath("/path");
@@ -445,7 +463,7 @@ public class FlinkExecutionEnvironmentsTest {
 
   @Test
   public void shouldFailOnNoStoragePathProvided() {
-    FlinkPipelineOptions options = FlinkPipelineOptions.defaults();
+    FlinkPipelineOptions options = getDefaultPipelineOptions();
     options.setStreaming(true);
     options.setStateBackend("unknown");
 
@@ -457,7 +475,7 @@ public class FlinkExecutionEnvironmentsTest {
 
   @Test
   public void shouldCreateFileSystemStateBackend() {
-    FlinkPipelineOptions options = FlinkPipelineOptions.defaults();
+    FlinkPipelineOptions options = getDefaultPipelineOptions();
     options.setStreaming(true);
     options.setStateBackend("fileSystem");
     options.setStateBackendStoragePath(temporaryFolder.getRoot().toURI().toString());
@@ -470,7 +488,7 @@ public class FlinkExecutionEnvironmentsTest {
 
   @Test
   public void shouldCreateRocksDbStateBackend() {
-    FlinkPipelineOptions options = FlinkPipelineOptions.defaults();
+    FlinkPipelineOptions options = getDefaultPipelineOptions();
     options.setStreaming(true);
     options.setStateBackend("rocksDB");
     options.setStateBackendStoragePath(temporaryFolder.getRoot().toURI().toString());