+++ runners/flink/src/test/java/org/apache/beam/runners/flink/FlinkStreamingTransformTranslatorsTest.java
@@ -29,8 +29,8 @@ import java.util.List;
 import java.util.Map;
 import org.apache.beam.runners.core.construction.PTransformTranslation;
 import org.apache.beam.runners.core.construction.SplittableParDo;
-import org.apache.beam.runners.flink.FlinkStreamingTransformTranslators.UnboundedSourceWrapperNoValueWithRecordId;
-import org.apache.beam.runners.flink.translation.wrappers.streaming.io.UnboundedSourceWrapper;
+import org.apache.beam.runners.flink.translation.wrappers.streaming.io.source.FlinkSource;
+import org.apache.beam.runners.flink.translation.wrappers.streaming.io.source.bounded.FlinkBoundedSource;
 import org.apache.beam.sdk.Pipeline;
 import org.apache.beam.sdk.coders.Coder;
 import org.apache.beam.sdk.coders.StringUtf8Coder;
@@ -49,8 +49,8 @@ import org.apache.beam.sdk.values.TupleTag;
 import org.apache.beam.sdk.values.WindowingStrategy;
 import org.apache.beam.vendor.guava.v32_1_2_jre.com.google.common.collect.Iterables;
 import org.apache.flink.streaming.api.environment.StreamExecutionEnvironment;
-import org.apache.flink.streaming.api.transformations.LegacySourceTransformation;
 import org.apache.flink.streaming.api.transformations.OneInputTransformation;
+import org.apache.flink.streaming.api.transformations.SourceTransformation;
 import org.checkerframework.checker.nullness.qual.Nullable;
 import org.junit.Test;
 
@@ -76,11 +76,10 @@ public class FlinkStreamingTransformTranslatorsTest {
     Object sourceTransform =
         applyReadSourceTransform(transform, PCollection.IsBounded.BOUNDED, env);
 
-    UnboundedSourceWrapperNoValueWithRecordId source =
-        (UnboundedSourceWrapperNoValueWithRecordId)
-            ((LegacySourceTransformation<?>) sourceTransform).getOperator().getUserFunction();
+    FlinkBoundedSource<?> source =
+        (FlinkBoundedSource<?>) ((SourceTransformation<?, ?, ?>) sourceTransform).getSource();
 
-    assertEquals(maxParallelism, source.getUnderlyingSource().getSplitSources().size());
+    assertEquals(maxParallelism, source.getNumSplits());
   }
 
   @Test
@@ -96,11 +95,10 @@ public class FlinkStreamingTransformTranslatorsTest {
     Object sourceTransform =
         applyReadSourceTransform(transform, PCollection.IsBounded.BOUNDED, env);
 
-    UnboundedSourceWrapperNoValueWithRecordId source =
-        (UnboundedSourceWrapperNoValueWithRecordId)
-            ((LegacySourceTransformation<?>) sourceTransform).getOperator().getUserFunction();
+    FlinkBoundedSource<?> source =
+        (FlinkBoundedSource<?>) ((SourceTransformation<?, ?, ?>) sourceTransform).getSource();
 
-    assertEquals(parallelism, source.getUnderlyingSource().getSplitSources().size());
+    assertEquals(parallelism, source.getNumSplits());
   }
 
   @Test
@@ -119,13 +117,12 @@ public class FlinkStreamingTransformTranslatorsTest {
         (OneInputTransformation)
             applyReadSourceTransform(transform, PCollection.IsBounded.UNBOUNDED, env);
 
-    UnboundedSourceWrapper source =
-        (UnboundedSourceWrapper)
-            ((LegacySourceTransformation<?>) Iterables.getOnlyElement(sourceTransform.getInputs()))
-                .getOperator()
-                .getUserFunction();
+    FlinkSource<?, ?> source =
+        (FlinkSource<?, ?>)
+            ((SourceTransformation<?, ?, ?>) Iterables.getOnlyElement(sourceTransform.getInputs()))
+                .getSource();
 
-    assertEquals(maxParallelism, source.getSplitSources().size());
+    assertEquals(maxParallelism, source.getNumSplits());
   }
 
   @Test
@@ -142,13 +139,12 @@ public class FlinkStreamingTransformTranslatorsTest {
         (OneInputTransformation)
             applyReadSourceTransform(transform, PCollection.IsBounded.UNBOUNDED, env);
 
-    UnboundedSourceWrapper source =
-        (UnboundedSourceWrapper)
-            ((LegacySourceTransformation<?>) Iterables.getOnlyElement(sourceTransform.getInputs()))
-                .getOperator()
-                .getUserFunction();
+    FlinkSource<?, ?> source =
+        (FlinkSource<?, ?>)
+            ((SourceTransformation<?, ?, ?>) Iterables.getOnlyElement(sourceTransform.getInputs()))
+                .getSource();
 
-    assertEquals(parallelism, source.getSplitSources().size());
+    assertEquals(parallelism, source.getNumSplits());
   }
 
   private Object applyReadSourceTransform(
@@ -157,7 +153,7 @@ public class FlinkStreamingTransformTranslatorsTest {
     FlinkStreamingPipelineTranslator.StreamTransformTranslator<PTransform<?, ?>> translator =
         getReadSourceTranslator();
     FlinkStreamingTranslationContext ctx =
-        new FlinkStreamingTranslationContext(env, PipelineOptionsFactory.create());
+        new FlinkStreamingTranslationContext(env, PipelineOptionsFactory.create(), true);
 
     Pipeline pipeline = Pipeline.create();
     PCollection<String> pc =